Finished refactoring coordinates usage

+++ src/interface_primitives/coloredRectangles.h
@@ -9,13 +9,13 @@
 class ColoredRectangles {
 public:
   ColoredRectangles (glm::vec4 color, bool filled = true);
-  ColoredRectangles (glm::vec4 color, const glm::vec4& rectangle, bool filled = true);
-  ColoredRectangles (glm::vec4 color, const std::vector<glm::vec4>& rectangles, bool filled = true);
+  ColoredRectangles (glm::vec4 color, const glm::ivec4& rectangle, bool filled = true);
+  ColoredRectangles (glm::vec4 color, const std::vector<glm::ivec4>& rectangles, bool filled = true);
 
   void bindShaderAndDraw() const;
 
-  void setRectangles(const glm::vec4& rectangle);
-  void setRectangles(const std::vector<glm::vec4>& rectangles);
+  void setRectangles(const glm::ivec4& rectangle);
+  void setRectangles(const std::vector<glm::ivec4>& rectangles);
 
   // Loaded in Interface init function to ensure it is only loaded once
   static void loadShader();
